+++ dev/examples/draggable.tsx
@@ -2,11 +2,19 @@ import * as React from "react"
 import { motion } from "../../src"
 
 const styleA = {
-    width: 300,
-    height: 300,
-    background: "blue",
+    width: 200,
+    height: 200,
+    background: "white",
+    borderRadius: 20,
 }
 
 export const App = () => {
-    return <motion.div drag style={styleA} />
+    return (
+        <motion.div
+            drag
+            whileTap={{ scale: 0.95 }}
+            whileHover={{ scale: 1.1 }}
+            style={styleA}
+        />
+    )
 }