+++ dev/styled.ts
@@ -3,7 +3,7 @@ import styled from "styled-components"
 export const Container = styled.div`
     min-width: 100vw;
     min-height: 100vh;
-    background: #081320;
+    background: #111;
     display: flex;
     align-items: center;
     justify-content: center;